+++ lib/honeybadger/notice.rb
@@ -319,14 +319,7 @@ module Honeybadger
     def extract_source_from_backtrace(backtrace, config, opts)
       return nil if backtrace.lines.empty?
 
-      # ActionView::Template::Error has its own source_extract method.
-      # If present, use that instead.
-      if exception.respond_to?(:source_extract) && exception.source_extract
-        Hash[exception.source_extract.split("\n").map do |line|
-          parts = line.split(': ')
-          [parts[0].strip, parts[1] || '']
-        end]
-      elsif backtrace.application_lines.any?
+      if backtrace.application_lines.any?
         backtrace.application_lines.first.source(config[:'exceptions.source_radius'])
       else
         backtrace.lines.first.source(config[:'exceptions.source_radius'])