+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Fixed
+- Squashed a bug where the wrong source extract was sent for some instances of
+  `ActionView::Template::Error`.
 
 ## [2.5.2] - 2016-03-08
 ### Fixed